在.net3.5环境下成功编译并运行，让还在使用vs2008的童靴们也能享受NanUI的强大和便利(^_^_^)

+++ NetDimension.NanUI/ChromiumFX/Source/CfxBinaryValue.cs
@@ -67,7 +67,7 @@ namespace Chromium {
 
             var po = new PinnedObject(buffer);
 
-            var retval = CfxApi.cfx_binary_value_get_data(NativePtr, po.PinnedPtr + bufferOffset, maxBytes, dataOffset);
+            var retval = CfxApi.cfx_binary_value_get_data(NativePtr, (IntPtr)((int)po.PinnedPtr + bufferOffset), maxBytes, dataOffset);
             po.Free();
             return retval;
         }
